Create Session_5_Scenario_1_FTP_com_2_File_Processin
I want to make some changes in the script : 

1- you use the `USERNAME` variable in your code , identical for both vm2 and vm3. the username may will be different, so define two separate variable for username of each vm.

2- to make condition simpler, use absolute path instead of relative path for source and destination. i mean the source path must be a list of path tha define in the dag file. for example : ftp_file_path = ['/home/rocky/in/sample1.txt', '/home/rocky/bon/in/sample3.txt', '/home/rocky/card/in/sample2.txt']. so i dont need task discover_files any more.   

3- i want to change the simple_file_manager.py python script that make change in the text fil as: 
the script must take the text file and generate two file with postfix .inv .dat format, (with the same name of the text file). imagine that the sample1.txt has 10 line:
A-  in .dat file, it have to be 10 Character randomly using `v` and `x` Character (related and represent to each line in sample1.txt). for example the content of the sample1.dat must be `vxvvvxxxvx`. 
B- and the content of sample1.inv must be the exact line (in sample1.txt) that related to `v` charector that generated in sample1.dat . 

4- the before the trensfering file it have to check that if the file exist in the destination path. if there was no file , it have to transfer. if the file exist in the destination path, (with the exactly identical size and name)File transfer should not be performed. if the file exist and the size of the file are different (comparision between files in source path and dest path) it have to notiff with proper message. and raise proper error that cause fail and the rest of the proccess Should not be done. 

5- there is a critical checking before running simple_file_manager.py script  : 
if in the destination all .txt and .dat and .inv exist , it DOSENT have to run script at all. not at all ! (because it mean that the script runned before and the .inv and .dat file already generated.) so define it as the new task. (or in a transfereing task , witch ever is better chose on you)


again consider that your code must be simple that be appropriate for new learner in python. and it have to be functional and reusable in each part of it.